refactor: unify toolbar button
